requirements.txt
# Core data science and backtesting dependencies
pandas>=1.5.0
numpy>=1.21.0

# Database
sqlite3  # Built-in Python library

# Technical analysis
pandas-ta>=0.3.14b

# Schwab API client
schwab-py>=1.0.0

# Optional: For enhanced data fetching and analysis
yfinance>=0.1.87  # Alternative data source
requests>=2.28.0  # HTTP requests

# Development and testing
pytest>=7.0.0
black>=22.0.0  # Code formatting
flake8>=4.0.0  # Linting

# Logging and utilities
python-dateutil>=2.8.0</content>
<parameter name="filePath">c:\Users\panic\OneDrive\Development\DxOrderDaemon\SchwabAlgo\requirements.txt